Punkt mit double Koordinaten

+++ src/forms/Point.java
@@ -6,8 +6,8 @@ package forms;
  */
 public class Point
 {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     /**
      * Constructor for creating Point objects
@@ -17,18 +17,18 @@ public class Point
         this(0, 0);
     }
 
-    public Point(int x, int y)
+    public Point(double x, double y)
     {
         this.x = x;
         this.y = y;
     }
 
-    public int getX()
+    public double getX()
     {
         return this.x;
     }
 
-    public int getY()
+    public double getY()
     {
         return this.y;
     }
@@ -52,6 +52,6 @@ public class Point
         xv = tx + center.getX();
         yv = ty + center.getY();
 
-        return new Point((int)xv, (int)yv);
+        return new Point(xv, yv);
     }
 }
